+++ sertit/__init__.py
@@ -20,4 +20,4 @@
 .. include:: ../README.md
 """
 
-__version__ = "1.7.0"
+__version__ = "1.7.1"